made changes necessitated by moving ObservationMetaData from sims_catalogs_generation to sims_utils
