+++ python/lsst/sims/catalogs/measures/instance/InstanceCatalog.py
@@ -5,7 +5,7 @@ import inspect
 import re
 import copy
 from .fileMaps import defaultSpecMap
-from lsst.sims.catalogs.generation.db import ObservationMetaData
+from lsst.sims.utils import ObservationMetaData
 
 __all__ = ["InstanceCatalog", "is_null"]
