+++ tests/testCatalogs.py
@@ -4,7 +4,8 @@ import numpy
 import unittest
 import eups
 import lsst.utils.tests as utilsTests
-from lsst.sims.catalogs.generation.db import CatalogDBObject, ObservationMetaData
+from lsst.sims.utils import ObservationMetaData
+from lsst.sims.catalogs.generation.db import CatalogDBObject
 from lsst.sims.catalogs.measures.instance import InstanceCatalog, compound
 import lsst.sims.catalogs.generation.utils.testUtils as tu
 from lsst.sims.utils import haversine
